feat: modify process for working with gpl3 license
Licensee doesn't detect an attribution in a GPL3 license, so this can only be used when working with
a BSD-3-Clause license.

Unfortunately, the license that's reported by the API is always what's in the main branch, so you
can't use this as a CI check; it will fail if you change your license on a branch. Thus, the checks
related to this were removed.