feat: convert action to nodejs
Extending the Bash code to support Cargo workspace repositories was getting out of hand, so I
decided to convert the action to Node. I should have probably written it this way to begin with, but
quite frankly I just forgot it was something you could do.

This first step is to verify the license files and to make sure they all use the same license. It
works for both a single-crate repository and a workspace. The verification of the source files
containing a copyright header will follow in another commit.