feat: verify source files have copyright notices
All the Rust source files in the repository (whether it's a workspace or not) will be checked for
the presence of a copyright notice using ripgrep (rg).

It was necessary to use `--fixed-strings` with rg, because by default, it will use regular
expressions, so there could be issues with escaping parentheses characters and such things. Easier
just to treat the search string as a literal string to search for. We also tell rg to return files
that *don't* contain matches for the string, so as a result, you get all the files that are missing
the copyright notice.

There is a comment in the diff that explains the important reason why the Github Actions `exec`
package could not be used for running rg.