fix: remove double quotes from attribution
When the attribution is read by Licensee, it will be encapsulated in double quotes. We need to
remove these from the string when it gets passed through the shell when we run ripgrep.

I've also changed the ripgrep process run to check the exit code, as we want the action to fail if
there's an error running the process. However, there is a caveat here. See the comment in the diff.