feat: support multiple attributions in license
In a forked repository, it's possible for there to be multiple attributions in the LICENSE file: the
original copyright holder and our own new attribution. Some licenses, e.g., MIT, require the
original to be preserved in the fork.

Multiple attributions will be returned from `jq.run` as a single string, with each attribution
delineated by a newline character, so we can use a split to determine if multiple are present. If
so, we will use the one that contains the company name specified in by our `company-name` workflow
input.